requirements.txt
chromalog
click
pillow
numpy